Initial: first working copy of imgobjsync
imgobjsync is a program that walks a directory tree. For each image file
encountered it:

	* resizes the image to fit within a specified pixel size
	* uploads both the original and resized image to DreamObjects

If a file cannot be resized or uploaded it's skipped and imgobjsync
continues to try the remaining files in the directory.

If both an image *and* its resized version already exist (by name) on
DreamObjects that image is skipped.

TODO:

	* Use goroutines to process images concurrently
	* don't crash when we don't include a directory on invocation
	* add an option to force re-uploading
	* Switch from minio-go to aws
	* Set object permissions to public-read
	* Support for multiple sizes?